% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{exampleEdges}
\alias{exampleEdges}
\title{A data file of edges from Regulatory networks.}
\format{A data set with 68 observations of 4 variables.}
\source{
See Mike Bostock \url{http://bl.ocks.org/mbostock/4062045}.
}
\description{
A data file of edges from Regulatory networks.
}
\keyword{datasets}

