% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{exampleHullColours}
\alias{exampleHullColours}
\title{A string of javascript D3 color scale for membership colors of exampleNodes}
\format{A D3 String}
\description{
A string of javascript D3 color scale for membership colors of exampleNodes
}
\keyword{datasets}

