% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{exampleNodeColours}
\alias{exampleNodeColours}
\title{A string of javascript D3 color scale for exampleNodes}
\format{A D3 string}
\description{
A string of javascript D3 color scale for exampleNodes
}
\keyword{datasets}

